refactor: validator UI 폴더로 이동
