+++ src/ts/ui/ProductManagementUI/ProductAdditionUI.js
@@ -1,5 +1,5 @@
 import { $ } from '../../utils/dom';
-import { validateProductInfo } from '../../utils/validator';
+import { validateProductInfo } from './validator';
 import { viewPainter } from '../ViewPainter';
 
 export default class ProductAdditionUI {
